+++ DbgCensus.Rest/Extensions/IServiceCollectionExtensions.cs
@@ -1,12 +1,12 @@
 ﻿using DbgCensus.Rest.Abstractions;
 using DbgCensus.Rest.Abstractions.Queries;
-using DbgCensus.Rest.Polly;
 using DbgCensus.Rest.Queries;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Options;
 using Polly;
 using Polly.Contrib.WaitAndRetry;
+using Polly.Extensions.Http;
 using System;
 using System.Net.Http;
 using System.Text.Json;
@@ -15,6 +15,8 @@ namespace DbgCensus.Rest.Extensions;
 
 public static class IServiceCollectionExtensions
 {
+    private static int _serviceIDIndex;
+
     /// <summary>
     /// Adds required services for interacting with the Census REST API.
     /// </summary>
@@ -29,24 +31,33 @@ public static class IServiceCollectionExtensions
         );
 
         serviceCollection.AddHttpClient<ICensusRestClient, CensusRestClient>()
-            .ConfigurePrimaryHttpMessageHandler(() => new HttpClientHandler {AllowAutoRedirect = false})
-            .AddTransientHttpErrorPolicy
-            (
-                builder => builder.WaitAndRetryAsync
-                (
-                    Backoff.DecorrelatedJitterBackoffV2(TimeSpan.FromSeconds(1), 4)
-                )
-            )
-            .AddTransientHttpErrorPolicy
-            (
-                builder => builder.CircuitBreakerAsync(4, TimeSpan.FromSeconds(30))
-            )
+            .ConfigurePrimaryHttpMessageHandler(() => new HttpClientHandler { AllowAutoRedirect = false })
             .AddPolicyHandler
             (
-                (s, b) => new ServiceIDRotationPolicy<HttpResponseMessage>
+                (services, _) => HttpPolicyExtensions.HandleTransientHttpError()
+                    .WaitAndRetryAsync
                     (
-                        s.GetRequiredService<IOptionsMonitor<CensusQueryOptions>>()
+                        Backoff.DecorrelatedJitterBackoffV2(TimeSpan.FromSeconds(1), 4),
+                        onRetry: (_, _, retryAttempt, _) =>
+                        {
+                            if (retryAttempt != 4)
+                                return;
+
+                            CensusQueryOptions qOptions = services.GetRequiredService<IOptionsMonitor<CensusQueryOptions>>().CurrentValue;
+                            if (qOptions.ServiceIDs.Count == 0)
+                                return;
+
+                            _serviceIDIndex++;
+                            if (_serviceIDIndex >= qOptions.ServiceIDs.Count)
+                                _serviceIDIndex = 0;
+
+                            qOptions.ServiceId = qOptions.ServiceIDs[_serviceIDIndex];
+                        }
                     )
+            )
+            .AddTransientHttpErrorPolicy
+            (
+                builder => builder.CircuitBreakerAsync(4, TimeSpan.FromSeconds(15))
             );
 
         serviceCollection.TryAddSingleton<IQueryBuilderFactory, QueryBuilderFactory>();